requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
Django==3.1.7
django-cors-headers==3.7.0
django-debug-toolbar==3.2
django-filter==2.4.0
django-redis==4.12.1
djangorestframework==3.12.2
djangorestframework-deep==0.1.1
drf-jwt==1.18.0
freezegun==1.1.0
idna==2.10
Markdown==3.3.4
mysqlclient==2.0.3
Pillow==8.1.2
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
